mos6502: Add some memory reading and writing abstractions
Add memRead() and memRead16() which return a byte and 2 bytes of
memory from the provided base address respectively.

Add memWrite() and memWrite16() which store a byte and 2 bytes into
memory at the provided base address respectively.

These abstractions should be useful later when we add mappers, but for
now are handy enough for the multi-byte reading and writing we're
beginning to add for memory addressing mode support.