mos6502: Add getOperandAddr to handle memory addressing modes
When we need to fetch an instruction, we need to handle the different
addressing modes. Implement that handling in a central method so we
can call it from all of the opFOO methods as we implement the next
instructions that rely on modes other than ACCUMULATOR and IMPLICIT.

Add some basic tests. We likely want to extend these later, but for
now, we have a single test for each mode that is actually interesting.