6502: Make the CPU require a mapper
This is the first step toward allowing a real NES cartridge ROM to be
used during program execution. We supply a dummy mapper for all tests
which simply hides a []uint8 slice of MEM_SIZE elements which is what
we had before, directly via an internal slice. This makes the tests
not depend on a ROM, but still exercises the mapper abstraction.