6502, Mappers: Change the memory interface
Leave local RAM access in a different memory object so it's not the
responsibility of the mapper anymore. This leaves mappers as smaller
logic shims around the NES ROM data where we can handle bank mapping,
etc. This is in prep for adding the PPU which also has 2k of VRAM and
doesn't share that via the mapper.