console: Rename machine to Bus
We're going to refactor how some of the components hang together to
allow (eg) the 6502 to be used in ways that don't matter if it's in an
NES or a general computer. IOW, we're going to abstract understanding
of memory areas out of the CPU code and into the Bus. This will let us
maintain better cpu tests more easily and should be a better structure
to the code as well.