Clean up the memory abstratction
Instead of passing a bus and a mapper to CPU, pass just the bus. To do
this, we request a Memory interface object, which the bus
implements. Do the same for the PPU as well, although we'll need to
make futher changes to what we pass in there so that it can use the
dedicated vram it contains.