Refactor PPU and Bus interactions
Allow the PPU to read ROM data via the bus through a ChrRead
method. Drop the write handling because it only ever writes its own
internal vram. Wire up a means to trigger the NMI in the cpu via the
bus. A few more minor refactorings to follow here so we completely
remove knowledge of the ROM code from the PPU.