MOS6502: Add better interrupt handling support
Add support for IRQ and change the mechanism used to signal interrupts
so that we can detect we aren't pending INT_NONE and then jump to the
value of cpu's pendingInterrupt, otherwise. This may need further
tweaking still as we figure out things like cycle timing which might
mean we need to handle multiple interrupts being fired simultaneously,
etc.