Bus: Make the Bus drive it's own Run() method
Instead of using the CPU's Run() method, let the bus have its
own. This means we don't need callback handling, etc. We'll remove
that from the CPU code later, but leave it for now as we're still
experimenting with how to wire this all up to manage the cycle counts
effectively.