gintendo: Use the ebiten.RunGame as the primary entrypoint now
Drop the goroutine used for ebiten and have the bus.Update() method
drive everything we'd have done from the BIOS() (typically via the 'r'
command). This makes ebiten the arbiter of everything. We now run a
rough estimate of ticks within the Update() that would be happening
every 1/60 of a second. This is a bit janky right now, but we'll clean
that up.