MOS6502: Handle wrapping in INDIRECT_{X,Y} addressing
Wrap addresses within the zero page for INDIRECT_X and INDIRECT_Y
lookups. This ensures nestest.nes will pass the INDIRECT_X
tests. INDIRECT_Y was passing already, but it a) still passes with
this test and b) does seem to be what the hardware does in this mode,
per documentation.